fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@12640 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/devices/MSDevice_Vehroutes.cpp
@@ -188,7 +188,7 @@ MSDevice_Vehroutes::writeXMLRoute(OutputDevice& os, int index) const {
             int remainingWithExitTime = (int)myExits.size() - numWritten;
             assert(remainingWithExitTime >= 0);
             assert(remainingWithExitTime <= (int)myCurrentRoute->size());
-            if (remainingWithExitTime < myCurrentRoute->size()) {
+            if (remainingWithExitTime < (int)myCurrentRoute->size()) {
                 upTo = *(myCurrentRoute->begin() + remainingWithExitTime);
             }
         }
